+++ packages/winnow/lib/normalize-query-options/output-data-spatial-reference.js
@@ -1,4 +1,4 @@
-const normalizeSpatialReference = require('../helpers/normalize-spatial-reference')
+const normalizeSpatialReference = require('./spatial-reference')
 const logWarning = process.env.NODE_ENV !== 'production' && process.env.KOOP_WARNINGS !== 'suppress'
 
 function normalizeOutputDataSpatialReference (options = {}) {
@@ -6,21 +6,20 @@ function normalizeOutputDataSpatialReference (options = {}) {
     srsname,
     srsName,
     projection,
-    outSR
+    outputCrs,
+    outSR,
+    inputCrs,
+    sourceSR
   } = options
 
-  const outputSpatialReference = projection || srsname || srsName || outSR
+  // if no output spatial reference set (outputCrs, projection, srsname, srsName, outSR), assume output will be same as input (inputCrs, sourceSR)
+  const outputSpatialReference = outputCrs || projection || srsname || srsName || outSR || inputCrs || sourceSR || 4326
 
-  if (!outputSpatialReference) return
   const spatialReference = normalizeSpatialReference(outputSpatialReference)
 
-  if (!spatialReference) {
-    if (logWarning) console.log(`WARNING: spatial reference "${outputSpatialReference}" could not be normalized. Defaulting to EPSG:4326.`)
-    return
-  }
+  if (logWarning) console.log(`WARNING: spatial reference "${outputSpatialReference}" could not be normalized. Defaulting to EPSG:4326.`)
 
-  const { wkid, wkt } = spatialReference
-  return wkid ? `EPSG:${wkid}` : wkt
+  return spatialReference || { wkid: 4326 }
 }
 
 module.exports = normalizeOutputDataSpatialReference