+++ packages/winnow/lib/normalize-query-options/source-data-spatial-reference.js
@@ -1,21 +1,17 @@
 const _ = require('lodash')
-const normalizeSpatialReference = require('../helpers/normalize-spatial-reference')
+const normalizeSpatialReference = require('./spatial-reference')
 const logWarning = process.env.NODE_ENV !== 'production' && process.env.KOOP_WARNINGS !== 'suppress'
 const OGC_WGS84 = 'ogc:1.3:crs84'
 
 function normalizeSourceDataSpatialReference ({ inputCrs, sourceSR, collection } = {}) {
   const sourceDataSpatialReference = inputCrs || sourceSR || getCollectionCrs(collection)
-  if (!sourceDataSpatialReference) return 'EPSG:4326'
+  if (!sourceDataSpatialReference) return { wkid: 4326 }
 
   const spatialReference = normalizeSpatialReference(sourceDataSpatialReference)
 
-  if (!spatialReference) {
-    if (logWarning) console.log(`WARNING: spatial reference "${sourceDataSpatialReference}" could not be normalized. Defaulting to EPSG:4326.`)
-    return 'EPSG:4326'
-  }
+  if (logWarning) console.log(`WARNING: spatial reference "${sourceDataSpatialReference}" could not be normalized. Defaulting to EPSG:4326.`)
 
-  const { wkid, wkt } = spatialReference
-  return wkid ? `EPSG:${wkid}` : wkt
+  return spatialReference || { wkid: 4326 }
 }
 
 function getCollectionCrs (collection) {