+++ packages/winnow/test/unit/normalize-query-options/geometry-filter-spatial-reference.spec.js
@@ -3,14 +3,14 @@ const normalizeGeometryFilterSpatialReference = require('../../../lib/normalize-
 
 test('normalize-query-options, geometry-filter-spatial-reference: undefined input', t => {
   t.plan(1)
-  const spatialRef = normalizeGeometryFilterSpatialReference()
-  t.equal(spatialRef, 'EPSG:4326')
+  const { wkid } = normalizeGeometryFilterSpatialReference()
+  t.equal(wkid, 4326)
 })
 
 test('normalize-query-options, geometry-filter-spatial-reference: undefined options', t => {
   t.plan(1)
-  const spatialRef = normalizeGeometryFilterSpatialReference({})
-  t.equal(spatialRef, 'EPSG:4326')
+  const { wkid } = normalizeGeometryFilterSpatialReference({})
+  t.equal(wkid, 4326)
 })
 
 test('normalize-query-options, geometry-filter-spatial-reference: defer to geometry option', t => {
@@ -20,8 +20,8 @@ test('normalize-query-options, geometry-filter-spatial-reference: defer to geome
     bbox: '100,200,300,400,4326',
     inSR: '4269'
   }
-  const spatialRef = normalizeGeometryFilterSpatialReference(options)
-  t.equal(spatialRef, 'EPSG:3857')
+  const { wkid } = normalizeGeometryFilterSpatialReference(options)
+  t.equal(wkid, 3857)
 })
 
 test('normalize-query-options, geometry-filter-spatial-reference: defer to bbox option if no geometry', t => {
@@ -30,8 +30,8 @@ test('normalize-query-options, geometry-filter-spatial-reference: defer to bbox 
     bbox: '100,200,300,400,3857',
     inSR: '4269'
   }
-  const spatialRef = normalizeGeometryFilterSpatialReference(options)
-  t.equal(spatialRef, 'EPSG:3857')
+  const { wkid } = normalizeGeometryFilterSpatialReference(options)
+  t.equal(wkid, 3857)
 })
 
 test('normalize-query-options, geometry-filter-spatial-reference: geometry filter bbox missing spatial reference', t => {
@@ -40,8 +40,8 @@ test('normalize-query-options, geometry-filter-spatial-reference: geometry filte
     geometry: '100,200,300,400',
     inSR: '4269'
   }
-  const spatialRef = normalizeGeometryFilterSpatialReference(options)
-  t.equal(spatialRef, 'EPSG:4269')
+  const { wkid } = normalizeGeometryFilterSpatialReference(options)
+  t.equal(wkid, 4269)
 })
 
 test('normalize-query-options, geometry-filter-spatial-reference: defer to geometry filter wkid', t => {
@@ -54,20 +54,20 @@ test('normalize-query-options, geometry-filter-spatial-reference: defer to geome
     },
     inSR: '4269'
   }
-  const spatialRef = normalizeGeometryFilterSpatialReference(options)
-  t.equal(spatialRef, 'EPSG:4326')
+  const { wkid } = normalizeGeometryFilterSpatialReference(options)
+  t.equal(wkid, 4326)
 })
 
 test('normalize-query-options, geometry-filter-spatial-reference: inSR string', t => {
   t.plan(1)
   const options = { inSR: '4269' }
-  const spatialRef = normalizeGeometryFilterSpatialReference(options)
-  t.equal(spatialRef, 'EPSG:4269')
+  const { wkid } = normalizeGeometryFilterSpatialReference(options)
+  t.equal(wkid, 4269)
 })
 
 test('normalize-query-options, geometry-filter-spatial-reference: inSR spatialReference object', t => {
   t.plan(1)
   const options = { inSR: { wkid: 4269 } }
-  const spatialRef = normalizeGeometryFilterSpatialReference(options)
-  t.equal(spatialRef, 'EPSG:4269')
+  const { wkid } = normalizeGeometryFilterSpatialReference(options)
+  t.equal(wkid, 4269)
 })